pyproject.toml


[project]
name = "open-lifeworlds-data-product-berlin-lor-daycare-centers"
version = "0.1.0"
description = "Data product combining Berlin LOR daycare centers data and geodata"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "open-lifeworlds-python-lib",
    "shapely>=2.1.1",
]

[tool.uv.sources]
open-lifeworlds-python-lib = { git = "https://github.com/open-lifeworlds/open-lifeworlds-python-lib.git" }
